requirements.txt
# Core dependencies
py-clob-client>=0.1.0
web3>=6.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
openai>=1.0.0

# Testing
hypothesis>=6.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Monitoring and observability
prometheus-client>=0.17.0
boto3>=1.28.0
watchtower>=3.0.0

# Utilities
requests>=2.31.0
colorama>=0.4.6
aiohttp>=3.9.0

# Rust bindings
maturin>=1.0.0
